simplified "scaling"
I figured that trying to scale these two items very precisely was too difficult and not the main objective of this project. So, instead of scaling the two pipes, we just made the pipes longer and move the center of the parent container up and down to make it look like the pipes are scaling with each other. The gap between the pipes is always the same which is perfect for predictability.