Created Main Menu Background + more
So I ran into a small bug while doing this. I learned that the startPos of the background does not change relative to the changing in size of the screen. This means if a player makes their screen larger while they are playing, then the background will not adjust to this and do its loop things inside the screen where the player can see it and it breaks the immersion. To fix this I had to go into the LeftAnchor.cs script and on like 24, I added the static start position of MoveBackground to equal the newPosition. I also had to reorganize the backgrounds to have the center of the game object to start at the edge of the screen so i didnt have to worry about an offset (offset is only for the bird)

So I had to do the same changes to the background of the gameplay as well to make the loop work on larger screens. There is the issue that if the players screen is like 100f wide then the looping might look off but I'm not gonna worry about that.

Also had an issue with the bird changing its z position from -1 to 0 when it hit the roof of the game, so I had to make sure that didnt happen in playerController.cs on line 47